pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.0",
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.1.0",
    "jq>=1.10.0",
    "langchain==0.3.27",
    "langchain-community>=0.3.31",
    "langchain-experimental>=0.3.4",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.35",
    "langchain-pinecone>=0.2.13",
    "langgraph>=1.0.1",
    "matplotlib>=3.10.8",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pdfminer>=20191125",
    "pymupdf>=1.26.7",
    "pypdf>=6.5.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=0.2.5.1",
    "streamlit>=1.52.2",
    "tiktoken>=0.12.0",
    "unstructured>=0.18.26",
    "wikipedia>=1.4.0",
]
